refactor: Cats and owners(service and controller) divided, now there are two controllers and services. Everything works!!!!
